Fixing up final rebase merge issues
